% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{armadillo.assume_role_with_web_identity}
\alias{armadillo.assume_role_with_web_identity}
\title{Retrieve temporary credentials from Armadillo MinIO filestore}
\usage{
armadillo.assume_role_with_web_identity(
  token,
  server,
  duration = 900,
  use = TRUE
)
}
\arguments{
\item{token}{ID token}

\item{server}{url of the MinIO server}

\item{duration}{duration in seconds}

\item{use}{boolean start using the credentials, default TRUE}
}
\value{
list of credentials, fit for use with \code{\link{set_credentials}}
}
\description{
Retrieve temporary credentials from Armadillo MinIO filestore
}
\keyword{internal}
