% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{armadillo.copy_resource}
\alias{armadillo.copy_resource}
\title{Copy resource}
\usage{
armadillo.copy_resource(
  project,
  folder,
  name,
  new_folder = folder,
  new_name = name
)
}
\arguments{
\item{project}{study or other variable collection}

\item{folder}{the folder containing the resource}

\item{name}{specific resource for copy action}

\item{new_folder}{name of the folder in which to place the copy, defaults to
folder}

\item{new_name}{name of the copy, defaults to name}
}
\value{
the response from the server
}
\description{
Copy resource
}
\examples{
\dontrun{
armadillo.copy_resource(
  project = "gecko",
  folder = "core_all",
  name = "table1",
  new_folder = "core_all_v2",
)
}

}
