% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{armadillo.copy_table}
\alias{armadillo.copy_table}
\title{Copy table}
\usage{
armadillo.copy_table(
  project,
  folder,
  name,
  new_project = project,
  new_folder = folder,
  new_name = name
)
}
\arguments{
\item{project}{study or other variable collection}

\item{folder}{the folder containing the table}

\item{name}{specific table for copy action}

\item{new_project}{new location of study or other variable collection}

\item{new_folder}{name of the folder in which to place the copy, defaults to
folder}

\item{new_name}{name of the copy, defaults to name}
}
\value{
the response from the server
}
\description{
Copy table
}
\examples{
\dontrun{
armadillo.copy_table(
  project = "gecko",
  folder = "core_all",
  name = "table1",
  new_project = "gecko",
  new_folder = "core_all_v2",
)
}

}
