% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{armadillo.copy_workspace}
\alias{armadillo.copy_workspace}
\title{Copy workspace}
\usage{
armadillo.copy_workspace(folder, name, new_folder, new_name = name)
}
\arguments{
\item{folder}{study or other variable collection}

\item{name}{specific workspace for copy action}

\item{new_folder}{new location of study or other variable collection}

\item{new_name}{name of the copy, defaults to name}
}
\description{
Copy workspace
}
\examples{
\dontrun{
armadillo.copy_worspace(
  folder = "gecko",
  name = "tim_subset_1",
  new_folder = "gecko_subset_1"
)
}

}
