% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{armadillo.create_project}
\alias{armadillo.create_project}
\title{Create a project for a variable collection}
\usage{
armadillo.create_project(
  project_name,
  users = NULL,
  overwrite_existing = FALSE
)
}
\arguments{
\item{project_name}{The name of the project to create. The project name
\itemize{
  \item{cannot be empty.}
  \item{must be no more than 56 characters.}
  \item{cannot end with a \code{-}.}
  \item{must consist of lowercase letters and numbers.}
  }}

\item{users}{A list collection
of the users that should have access to the project}

\item{overwrite_existing}{Boolean whenever a project
with the same name should be overwritten or not}
}
\description{
Create a project for a variable collection
}
\examples{
\dontrun{
armadillo.create_project("gecko")
}

}
