% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{armadillo.create_project_with_users}
\alias{armadillo.create_project_with_users}
\title{Create a project for a variable collection with supplied users}
\usage{
armadillo.create_project_with_users(project_name, users)
}
\arguments{
\item{project_name}{The name of the project to create. The project name
\itemize{
  \item{cannot be empty.}
  \item{must be no more than 56 characters.}
  \item{cannot end with a \code{-}.}
  \item{must consist of lowercase letters and numbers.}
  }}

\item{users}{A list collection of the users that should have access to the project}
}
\description{
Create a project for a variable collection with supplied users
}
\examples{
\dontrun{
armadillo.create_project("gecko", list("user1@users.com", "user2@users.com"))
}

}
