% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{armadillo.create_workspace}
\alias{armadillo.create_workspace}
\title{Combines tables in an .Rdata file and uploads them to a folder.}
\usage{
armadillo.create_workspace(folder, name, ...)
}
\arguments{
\item{folder}{the folder to upload to}

\item{name}{name of the workspace to create}

\item{...}{tables to upload to the dataset}
}
\description{
Combines tables in an .Rdata file and uploads them to a folder.
}
\examples{
\dontrun{
armadillo.create_workspace(
  folder = "gecko",
  name = "johan_subset_1",
  table1,
  table2
)
}

}
