% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder.R
\name{armadillo.delete_folder}
\alias{armadillo.delete_folder}
\title{Delete folder}
\usage{
armadillo.delete_folder(folder_name)
}
\arguments{
\item{folder_name}{study or collection of variables name}
}
\description{
A folder represents usually a study or collection of variables
}
\examples{
\dontrun{
armadillo.delete_folder(folder_name = "gecko")
}

}
