% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{armadillo.delete_project}
\alias{armadillo.delete_project}
\title{Delete project}
\usage{
armadillo.delete_project(project_name)
}
\arguments{
\item{project_name}{the name of the study or collection of variables name}
}
\description{
A project represents usually a study or collection of variables
}
\examples{
\dontrun{
armadillo.delete_project(project_name = "gecko")
}

}
