% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{armadillo.delete_project_folder}
\alias{armadillo.delete_project_folder}
\title{Delete project folder}
\usage{
armadillo.delete_project_folder(project, folder)
}
\arguments{
\item{project}{project to delete the object from}

\item{folder}{folder to delete the object from}
}
\description{
Delete project folder
}
