% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{armadillo.delete_resource}
\alias{armadillo.delete_resource}
\title{Delete resource}
\usage{
armadillo.delete_resource(project, folder, name)
}
\arguments{
\item{project}{project to delete the resource from}

\item{folder}{folder to delete the resource from}

\item{name}{resource name}
}
\description{
Delete resource
}
\examples{
\dontrun{
armadillo.delete_resource(
  project = "gecko",
  folder = "core_all",
  name = "table1"
)
}

}
