% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{armadillo.delete_table}
\alias{armadillo.delete_table}
\title{Delete table}
\usage{
armadillo.delete_table(project, folder, name)
}
\arguments{
\item{project}{project to delete the table from}

\item{folder}{folder to delete the table from}

\item{name}{table name}
}
\value{
TRUE if successful, otherwise an object of class aws_error details
}
\description{
Delete table
}
\examples{
\dontrun{
armadillo.delete_table(
  project = "gecko",
  folder = "core_all",
  name = "table1"
)
}

}
