% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder.R
\name{armadillo.delete_user_folder}
\alias{armadillo.delete_user_folder}
\title{Deletes a user's folder from the Users bucket}
\usage{
armadillo.delete_user_folder(user_name)
}
\arguments{
\item{user_name}{the folder of the user to delete}
}
\description{
Deletes a user's folder from the Users bucket
}
\examples{
\dontrun{
armadillo.delete_user_folder(user_name = "sido")
}

}
