% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{armadillo.delete_workspace}
\alias{armadillo.delete_workspace}
\title{Delete workspace}
\usage{
armadillo.delete_workspace(folder, name)
}
\arguments{
\item{folder}{folder to delete the workspace from}

\item{name}{workspace name}
}
\value{
TRUE if successful, otherwise an object of class aws_error details
if not.
}
\description{
Delete workspace
}
\examples{
\dontrun{
armadillo.delete_workspace(
  folder = "gecko",
  name = "johan_subset_1"
)
}

}
