% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{armadillo.get_project_users}
\alias{armadillo.get_project_users}
\title{Gets the users of an given project name}
\usage{
armadillo.get_project_users(project_name)
}
\arguments{
\item{project_name}{the name of the project to extract the users from}
}
\value{
List of all users within "project_name"
}
\description{
Gets the users of an given project name
}
\examples{
\dontrun{
armadillo.get_project_users("some-project")
}

}
