% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{armadillo.get_projects_info}
\alias{armadillo.get_projects_info}
\title{Gets the Projects information}
\usage{
armadillo.get_projects_info()
}
\value{
the projects and their information
}
\description{
Gets the Projects information
}
\examples{
\dontrun{
armadillo.get_projects_info()
}

}
