% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{armadillo.install_package}
\alias{armadillo.install_package}
\title{Install package}
\usage{
armadillo.install_package(path, profile, url)
}
\arguments{
\item{path}{the path to the package}

\item{profile}{the selected profile}

\item{url}{the url of the armadillo service}
}
\description{
Installs a user defined package.
}
