% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{armadillo.install_packages}
\alias{armadillo.install_packages}
\title{Install package}
\usage{
armadillo.install_packages(paths, profile = "default")
}
\arguments{
\item{paths}{the path to the package, can be a vector of string}

\item{profile}{the selected profile}
}
\description{
Installs a user defined package.
}
