% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{armadillo.list_projects}
\alias{armadillo.list_projects}
\title{List the projects}
\usage{
armadillo.list_projects()
}
\value{
the projects
}
\description{
List the projects
}
\examples{
\dontrun{
armadillo.list_projects()
}

}
