% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{armadillo.list_resources}
\alias{armadillo.list_resources}
\title{List the resources in a project}
\usage{
armadillo.list_resources(project)
}
\arguments{
\item{project}{the shared project in which the resources are located}
}
\value{
the resources in the project
}
\description{
List the resources in a project
}
\examples{
\dontrun{
armadillo.list_resources("gecko")
}

}
