% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder.R
\name{armadillo.list_shared_folders}
\alias{armadillo.list_shared_folders}
\title{List the shared folders}
\usage{
armadillo.list_shared_folders()
}
\value{
list of shared folders
}
\description{
List the shared folders
}
\examples{
\dontrun{
armadillo.list_shared_folders()
}

}
