% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{armadillo.list_tables}
\alias{armadillo.list_tables}
\title{List the tables in a project}
\usage{
armadillo.list_tables(project)
}
\arguments{
\item{project}{the shared project in which the tables are located}
}
\value{
the table names, without the extension
}
\description{
List the tables in a project
}
\examples{
\dontrun{
armadillo.list_tables("gecko")
}

}
