% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder.R
\name{armadillo.list_user_folders}
\alias{armadillo.list_user_folders}
\title{List the user folders}
\usage{
armadillo.list_user_folders()
}
\value{
list of user folders
}
\description{
List the user folders
}
\examples{
\dontrun{
armadillo.list_user_folders()
}

}
