% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{armadillo.list_workspaces}
\alias{armadillo.list_workspaces}
\title{List the workspaces}
\usage{
armadillo.list_workspaces(folder)
}
\arguments{
\item{folder}{the folder in which the workspaces are located}
}
\description{
List the workspaces
}
\examples{
\dontrun{
armadillo.list_workspaces(
  folder = "gecko"
)
}

}
