% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{armadillo.load_resource}
\alias{armadillo.load_resource}
\title{Load a resource from a project}
\usage{
armadillo.load_resource(project, folder, name, env = parent.frame())
}
\arguments{
\item{project}{study or collection variables}

\item{folder}{the folder containing the resource}

\item{name}{name of the resource}

\item{env}{The environment in which you want to load the resource.
Default is the parent.frame() from which the function is called.}
}
\value{
NULL, invisibly
}
\description{
Load a resource from a project
}
\examples{
\dontrun{
armadillo.load_resource(
  project = "gecko",
  folder = "core_all",
  name = "lc_core_1"
)

armadillo.load_resource(
  project = "gecko",
  folder = "core_all",
  name = "lc_core_1",
  env = globalenv()
)
}

}
