% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{armadillo.load_table}
\alias{armadillo.load_table}
\title{Load table based upon study project and tableset}
\usage{
armadillo.load_table(project, folder, name, env = parent.frame())
}
\arguments{
\item{project}{study or collection variables}

\item{folder}{the folder containing the table}

\item{name}{name of the table}

\item{env}{The environment in which you want to load the table.
Default is the parent.frame() from which the function is called.}
}
\value{
NULL, invisibly
}
\description{
Load table based upon study project and tableset
}
\examples{
\dontrun{
armadillo.load_table(
  project = "gecko",
  folder = "core_all",
  name = "lc_core_1"
)

armadillo.load_table(
  project = "gecko",
  folder = "core_all",
  name = "lc_core_1",
  env = globalenv()
)
}

}
