% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{armadillo.login_basic}
\alias{armadillo.login_basic}
\title{Login with username / password (meant for dev and test environments)}
\usage{
armadillo.login_basic(armadillo, username, password)
}
\arguments{
\item{armadillo}{URL of the Armadillo server}

\item{username}{the username}

\item{password}{the password}
}
\description{
Login with username / password (meant for dev and test environments)
}
\examples{
\dontrun{
armadillo.login(
  "https://armadillo.dev.molgenis.org", "admin", "admin"
)
armadillo.login("http://localhost:8080", "admin", "admin")
}

}
