% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{armadillo.make_views}
\alias{armadillo.make_views}
\title{Describes data available to subset and makes subset}
\usage{
armadillo.make_views(
  subset_def = NULL,
  source_project = NULL,
  source_folder = NULL,
  source_table = NULL,
  target_project = NULL,
  target_folder = NULL,
  target_table = NULL,
  target_variables = NULL,
  new_project = NULL,
  dry_run = NULL
)
}
\arguments{
\item{subset_def}{R object containing subset definition created by
\code{armadillo.subset_definition()}}

\item{source_project}{project from which to subset data}

\item{source_folder}{folder from which to subset data}

\item{source_table}{table from which to subset data}

\item{target_project}{project to upload subset to. Will be created if it doesn't exist}

\item{target_folder}{folder to upload subset to. Will be created if it doesn't exist}

\item{target_table}{table to upload subset to.}

\item{target_variables}{variables from `source_table` to include in the view}

\item{new_project}{Defunct: project to upload subset to. Replaced by `target_project`}

\item{dry_run}{Defunct: previously enabgled dry-run to check which variables are missing}
}
\value{
missing variables provided in the subset definition
}
\description{
This automates the process of:
\enumerate{
 \item Checking what data is available to create subsets
 \item Make the subset
}
}
\examples{
\dontrun{
armadillo.subset(
  source_project = "gecko",
  new_project = "study1",
  subset_def = local_subset
)
}

}
