% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armadillo-make-subset.R
\name{armadillo.makeSubset}
\alias{armadillo.makeSubset}
\title{Describes data available to subset and makes subset}
\usage{
armadillo.makeSubset(
  master_bucket = "alspac",
  core_folder = "2_1_core_1_0",
  outcome_folder = "1_1_outcome_1_0",
  ref_csv,
  type = c("describe", "subset", "both"),
  subset_name = NULL
)
}
\arguments{
\item{master_bucket}{directory from which to subset data}

\item{core_folder}{path of folder containing core data}

\item{outcome_folder}{path of folder containing outcome data}

\item{ref_csv}{csv file detailing the variables required 
in the subset. It must contain three columns:

1. type (either "core" or "outcome")
2. table (if type == "core", either "non_rep", "monthly_rep", "yearly_rep"
          of "trimester". if type == outcome, either "non_rep", "weekly_rep", 
          "monthly_rep" or "yearly_rep"))
3. variable (corresponding to LifeCycle variable name)}

\item{type}{Either 'describe' to describe data not available, 'subset' to subset
the data, (iii) 'both' to describe and subset the data}
}
\value{
if type == 'describe' or 'both' returns a tibble detailing variables not
available for subsetting
}
\description{
This automates the process of (i) checking what data is available to create
subsets, and (ii) makes the subsets.
}
