% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{armadillo.move_resource}
\alias{armadillo.move_resource}
\title{Move the resource}
\usage{
armadillo.move_resource(
  project,
  folder,
  name,
  new_folder = folder,
  new_name = name
)
}
\arguments{
\item{project}{a study or collection of variables}

\item{folder}{the folder containing the resource to move}

\item{name}{a resource to move}

\item{new_folder}{the folder to move the resource to, defaults to folder}

\item{new_name}{use to rename the file, defaults to name}
}
\value{
NULL, invisibly
}
\description{
Move the resource
}
\examples{
\dontrun{
armadillo.move_resource(
  project = "gecko",
  folder = "core_all",
  name = "table1",
  new_folder = "core_all_v2",
)
}

}
