% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{armadillo.move_table}
\alias{armadillo.move_table}
\title{Move the table}
\usage{
armadillo.move_table(
  project,
  folder,
  name,
  new_project,
  new_folder,
  new_name = name
)
}
\arguments{
\item{project}{a study or collection of variables}

\item{name}{a tableset to move}

\item{new_project}{a subset of the studies new location}
}
\description{
Move the table
}
\examples{
\dontrun{
armadillo.move_folder(
  project = "gecko",
  folder = "core_all",
  name = "table1",
  new_project = "gecko",
  new_folder = "core_all_v2",
)
}

}
