% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newSubsetDefinition.R
\name{armadillo.newSubsetDefinition}
\alias{armadillo.newSubsetDefinition}
\title{Builds an R object containing info required to make subsets}
\usage{
armadillo.newSubsetDefinition(vars = NULL, metadata = NULL)
}
\arguments{
\item{vars}{.csv file containing vars to subset. This file must contain
three columns with the headers 'folder', 'table' & 'variables'. 'Folder' must
refer to a folder in the armadillo project to be subsetted. 'Table' must refer
to a table within that folder. 'variables' must refer to variables within that
that table.}

\item{metadata}{.csv file containing meta variables to subset (optional)}
}
\description{
Builds an R object containing info required to make subsets
}
