% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{armadillo.set_credentials}
\alias{armadillo.set_credentials}
\title{Set credentials}
\usage{
armadillo.set_credentials(
  server,
  access_key,
  secret_key,
  session_token = NULL,
  use = TRUE
)
}
\arguments{
\item{server}{url of the MinIO server}

\item{access_key}{the MinIO access key}

\item{secret_key}{the MinIO secret key}

\item{session_token}{optional MinIO session token, default NULL}

\item{use}{boolean start using the credentials, default TRUE}
}
\value{
list of credentials, fit for use with \code{\link{set_credentials}}
}
\description{
Set credentials
}
\keyword{internal}
