% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{armadillo.subset}
\alias{armadillo.subset}
\title{Describes data available to subset and makes subset}
\usage{
armadillo.subset(
  source_project = NULL,
  new_project = NULL,
  subset_def = NULL,
  type = NULL,
  include_meta = NULL
)
}
\arguments{
\item{source_project}{project from which to subset data}

\item{new_project}{project to upload subset to. Will be created if it doesn't exist.}

\item{subset_def}{R object containing subset definition created by newSubsetDefinition()}

\item{type}{Either 'describe' to describe data not available, 'subset' to subset
the data, (iii) 'both' to describe and subset the data}

\item{include_meta}{wheter or not to automatically include key variables (defined in subset_def)}
}
\value{
if type == 'describe' or 'both' returns a tibble detailing variables not
available for subsetting
}
\description{
This automates the process of (i) checking what data is available to create
subsets, and (ii) makes the subsets.
}
