% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{armadillo.subset_definition}
\alias{armadillo.subset_definition}
\title{Builds an R object containing info required to make subsets}
\usage{
armadillo.subset_definition(vars = NULL, metadata = NULL)
}
\arguments{
\item{vars}{\code{.csv} file containing vars to subset.}

\item{metadata}{\code{.csv} file containing meta variables to subset (optional)}
}
\value{
a dataframe containing metadata and variables that is used for input in the \code{armadillo.subset()} method
}
\description{
This file must contain three columns with the headers 'folder', 'table' and
'variables'. 'Folder' must refer to a folder in the armadillo project to be
subsetted. 'Table' must refer to a table within that folder. 'variables' must
refer to variables within that table.
}
\examples{
\dontrun{
armadillo.subset_definition(
  vars = "C:/tmp/vars.csv",
  metadata = "C:/tmp/metadata.csv"
)
}

}
