% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{armadillo.upload_table}
\alias{armadillo.upload_table}
\title{Uploads a table to a folder in a project}
\usage{
armadillo.upload_table(project, folder, table, name = NULL)
}
\arguments{
\item{project}{the project to upload to}

\item{folder}{the folder to upload to}

\item{table}{the table to upload}

\item{name}{name of the table (optional)}
}
\value{
TRUE if successful, otherwise an object of class aws_error details
}
\description{
Uploads a table to a folder in a project
}
\examples{
\dontrun{
armadillo.upload_table(
  project = "gecko",
  folder = "core_all",
  table1
)
}

}
