% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{armadillo.whitelist_packages}
\alias{armadillo.whitelist_packages}
\title{Whitelist packages}
\usage{
armadillo.whitelist_packages(pkgs, profile = "default")
}
\arguments{
\item{pkgs}{packages that need to be added to the whitelist}

\item{profile}{the profile you want to whitelist the packages on}
}
\description{
When you installed developer packages you need to whitelist
them in other to be able to test them in the Armadillo platform.
}
