% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{assume_role_with_web_identity}
\alias{assume_role_with_web_identity}
\title{Login to the Armadillo MinIO filestore}
\usage{
assume_role_with_web_identity(
  token,
  server = Sys.getenv("AWS_S3_ENDPOINT"),
  duration = 900,
  use = TRUE
)
}
\arguments{
\item{token}{oAuth ID token}

\item{server}{url of the MinIO server}

\item{duration}{duration in seconds}

\item{use}{boolean indicating if the credentials should be set}
}
\description{
Login to the Armadillo MinIO filestore
}
