% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{copy_workspace}
\alias{copy_workspace}
\title{Copy workspace}
\usage{
copy_workspace(folder, name, new_folder, ...)
}
\arguments{
\item{folder}{study or other variable collection}

\item{name}{specific workspace for copy action}

\item{new_folder}{new location of study or other variable collection}

\item{...}{Additional arguments passed to \code{\link{s3HTTP}}}
}
\description{
Copy workspace
}
\examples{
\dontrun{
copy_worspace(
  folder = "gecko",
  name = "tim_subset_1",
  new_folder = "gecko_subset_1"
)
}

}
