% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder.R
\name{create_folder}
\alias{create_folder}
\title{Create a folder for a variable collection}
\usage{
create_folder(folder_name)
}
\arguments{
\item{folder_name}{name of the folder to create variable collection}
}
\description{
Create a folder for a variable collection
}
\examples{
\dontrun{
create_folder(folder_name = "gecko")
}

}
