% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder.R
\name{delete_user_folder}
\alias{delete_user_folder}
\title{Deletes a user's folder from the Users bucket}
\usage{
delete_user_folder(user_name, ...)
}
\arguments{
\item{user_name}{the folder of the user to delete}

\item{...}{Additional arguments passed to \code{\link{s3HTTP}}}
}
\description{
Deletes a user's folder from the Users bucket
}
\examples{
\dontrun{
delete_user_folder(
  user_name = "sido"
)
}

}
