% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{delete_workspace}
\alias{delete_workspace}
\title{Delete workspace}
\usage{
delete_workspace(folder, name)
}
\arguments{
\item{folder}{folder to delete the workspace from}

\item{name}{workspace name}
}
\description{
Delete workspace
}
\examples{
\dontrun{
delete_workspace(
  folder = "gecko",
  name = "johan_subset_1"
)
}

}
