% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{.add_slash_if_missing}
\alias{.add_slash_if_missing}
\title{Add a Slash to a URL if Missing}
\usage{
.add_slash_if_missing(url)
}
\arguments{
\item{url}{A character string representing the URL.}
}
\value{
A character string with a trailing slash if it was missing.
}
\description{
This function ensures that a given URL string ends with a slash ('/'). 
If the URL does not end with a slash, it appends one.
}
