% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{.check_folder_name}
\alias{.check_folder_name}
\title{Foldername check}
\usage{
.check_folder_name(name)
}
\arguments{
\item{name}{foldername}
}
\description{
For empty and can not contain underscores and dashes
}
\keyword{internal}
