% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_if_bucket_exists}
\alias{.check_if_bucket_exists}
\title{Check if bucket / folder exists}
\usage{
.check_if_bucket_exists(bucket_name, ...)
}
\arguments{
\item{bucket_name}{name of the folder}

\item{...}{Additional arguments passed to \code{\link{s3HTTP}}}
}
\description{
Check if bucket / folder exists
}
\keyword{internal}
