% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_if_workspace_exists}
\alias{.check_if_workspace_exists}
\title{Check if workspace exists}
\usage{
.check_if_workspace_exists(bucket_name, workspace_name, ...)
}
\arguments{
\item{bucket_name}{folder name}

\item{workspace_name}{workspace name}
}
\description{
Check if workspace exists
}
\keyword{internal}
