% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{.check_workspace_name}
\alias{.check_workspace_name}
\title{Check the workspace name}
\usage{
.check_workspace_name(name)
}
\arguments{
\item{name}{workspace name}
}
\description{
Name can not be empty and may not contain a slash
}
\keyword{internal}
