% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{.compress_resource}
\alias{.compress_resource}
\title{Helper function for compressing to an RDS file}
\usage{
.compress_resource(resource, file)
}
\arguments{
\item{resource}{the resource to write to file}

\item{file}{the name of the file (without extension)}
}
\value{
the extension of the file
}
\description{
Helper function for compressing to an RDS file
}
