% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{.compress_table}
\alias{.compress_table}
\title{Helper function for compressing to a parquet file}
\usage{
.compress_table(table, file)
}
\arguments{
\item{table}{the table to write to file}

\item{file}{the name of the file (without extension)}
}
\value{
the extension of the file
}
\description{
Helper function for compressing to a parquet file
}
