% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{.copy_object}
\alias{.copy_object}
\title{Copy object}
\usage{
.copy_object(
  project,
  folder,
  name,
  new_project = project,
  new_folder = folder,
  new_name = name,
  extension
)
}
\arguments{
\item{project}{study or other variable collection}

\item{folder}{the folder containing the object}

\item{name}{specific object for copy action}

\item{new_project}{new location of study or other variable collection}

\item{new_folder}{name of the folder in which to place the copy, defaults to
folder}

\item{new_name}{name of the copy, defaults to name}

\item{extension}{extension of the objects}
}
\description{
Copy object
}
