% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{.delete_object}
\alias{.delete_object}
\title{Delete object}
\usage{
.delete_object(project, folder, name, extension)
}
\arguments{
\item{project}{project to delete the object from}

\item{folder}{folder to delete the object from}

\item{name}{object name}

\item{extension}{extension of the object to delete}
}
\value{
TRUE if successful, otherwise an object of class aws_error details
if not.
}
\description{
Delete object
}
