% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{.get_linkfile_content}
\alias{.get_linkfile_content}
\title{Helper function to get the contents of a linkfile}
\usage{
.get_linkfile_content(project, object_name)
}
\arguments{
\item{project}{projectname where the linkfile is stored}

\item{object_name}{folder/name of linkfile}
}
\value{
the contents of the linkfile
}
\description{
Helper function to get the contents of a linkfile
}
