% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{.list_objects_by_extension}
\alias{.list_objects_by_extension}
\title{List the objects in a project based on the file type}
\usage{
.list_objects_by_extension(project, extension)
}
\arguments{
\item{project}{the shared project in which the objects are located}

\item{extension}{the extension of the files to list}
}
\description{
List the objects in a project based on the file type
}
