% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{.load_linked_table}
\alias{.load_linked_table}
\title{Helper function to extract the source parquet file in a linkfile}
\usage{
.load_linked_table(file, columns)
}
\arguments{
\item{file}{source table parquet file}

\item{columns}{character list of columns to select from source file}
}
\value{
the contents of the file, as data frame
}
\description{
Helper function to extract the source parquet file in a linkfile
}
