% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{.load_object}
\alias{.load_object}
\title{Load an object from a project}
\usage{
.load_object(project, folder, name, env, load_function, extension)
}
\arguments{
\item{project}{study or collection variables}

\item{folder}{the folder containing the object}

\item{name}{name of the object}

\item{env}{The environment in which you want to load the object.}

\item{load_function}{a function to extract the object with}

\item{extension}{the extension of the file}
}
\value{
NULL, invisibly
}
\description{
Load an object from a project
}
