% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{.load_resource}
\alias{.load_resource}
\title{Helper function to extract an RDS file}
\usage{
.load_resource(file)
}
\arguments{
\item{file}{file to extract}
}
\value{
the contents of the file
}
\description{
Helper function to extract an RDS file
}
