% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{.load_table}
\alias{.load_table}
\title{Helper function to extract a parquet file}
\usage{
.load_table(file)
}
\arguments{
\item{file}{file to extract}
}
\value{
the contents of the file, as data frame
}
\description{
Helper function to extract a parquet file
}
