% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{.move_object}
\alias{.move_object}
\title{Move the object}
\usage{
.move_object(
  project,
  folder,
  name,
  new_project,
  new_folder,
  new_name,
  extension
)
}
\arguments{
\item{project}{a study or collection of variables}

\item{folder}{the folder containing the object to move}

\item{name}{name of the object to move}

\item{new_project}{the project to move the object to}

\item{new_folder}{the folder to move the object to, defaults to folder}

\item{new_name}{use to rename the file, defaults to name}

\item{extension}{extension of the file}
}
\description{
Move the object
}
