% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{.to_readable_name}
\alias{.to_readable_name}
\title{Change bucketname to readable bucket name}
\usage{
.to_readable_name(bucket_name)
}
\arguments{
\item{bucket_name}{folder name}
}
\description{
Change bucketname to readable bucket name
}
\keyword{internal}
