% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{.to_shared_bucket_name}
\alias{.to_shared_bucket_name}
\title{Add 'shared-' prefix}
\usage{
.to_shared_bucket_name(folder_name)
}
\arguments{
\item{folder_name}{shared foldername}
}
\description{
Add 'shared-' prefix
}
\keyword{internal}
