% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{.to_user_bucket_name}
\alias{.to_user_bucket_name}
\title{Add 'user-' prefix}
\usage{
.to_user_bucket_name(user_name)
}
\arguments{
\item{folder_name}{user foldername}
}
\description{
Add 'user-' prefix
}
\keyword{internal}
