% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{.upload_object}
\alias{.upload_object}
\title{Uploads an object to a folder in a project}
\usage{
.upload_object(project, folder, object, name, compression_function)
}
\arguments{
\item{project}{the project to upload to}

\item{folder}{the folder to upload to}

\item{name}{name of the object}

\item{compression_function}{a function that compresses an object to a file}

\item{table}{the object to upload}
}
\value{
TRUE if successful, otherwise an object of class aws_error details
if not.
}
\description{
Uploads an object to a folder in a project
}
