% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_token}
\alias{get_token}
\title{Get ID Token}
\usage{
get_token(server)
}
\arguments{
\item{server}{the URL of the Armadillo server}
}
\value{
The ID token string
}
\description{
Get an ID token to log in on an Armadillo server or its MinIO filestore.
}
