% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder.R
\name{list_user_folders}
\alias{list_user_folders}
\title{List the user folders}
\usage{
list_user_folders(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link{s3HTTP}}}
}
\value{
list of user folders
}
\description{
List the user folders
}
\examples{
\dontrun{
list_user_folders()
}

}
