% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{login}
\alias{login}
\title{Login to the Armadillo MinIO filestore}
\usage{
login(server, token, duration = 3600)
}
\arguments{
\item{server}{url of the MinIO server}

\item{token}{oAuth ID token}

\item{duration}{duration in seconds}
}
\description{
Login to the Armadillo MinIO filestore
}
