% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolgenisArmadillo-defunct.R, R/auth.R
\name{MolgenisArmadillo-defunct}
\alias{MolgenisArmadillo-defunct}
\alias{armadillo.get_token}
\title{Defunct functions in package \pkg{MolgenisArmadillo}.}
\usage{
armadillo.get_token(server)
}
\description{
The functions listed below are defunct and no longer available.
When possible, alternative functions with similar functionality are
mentioned. Help pages for deprecated functions are available at
\code{help("<function>-defunct")}.
}
\section{\code{armadillo.get_token}}{

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

For \code{armadillo.get_token}, use
\code{DSMolgenisArmadillo::armadillo.get_credentials()} instead.
}

\keyword{internal}
