% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolgenisArmadillo-package.R
\docType{package}
\name{MolgenisArmadillo-package}
\alias{MolgenisArmadillo}
\alias{MolgenisArmadillo-package}
\title{MolgenisArmadillo: Armadillo client for the Armadillo service}
\description{
An Armadillo client for the Armadillo DataSHIELD service. Part of the MOLGENIS suite.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/molgenis/molgenis-r-armadillo}
  \item \url{https://molgenis.github.io/molgenis-r-armadillo}
  \item Report bugs at \url{https://github.com/molgenis/molgenis-r-armadillo/issues}
}

}
\author{
\strong{Maintainer}: Sido Haakma \email{s.haakma@rug.nl}

Authors:
\itemize{
  \item Tommy de Boer \email{t.de.boer01@umcg.nl}
  \item Fleur Kelpin \email{f.kelpin@umcg.nl}
}

}
\keyword{internal}
