% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{move_workspace}
\alias{move_workspace}
\title{Move the workspace}
\usage{
move_workspace(folder, name, new_folder, ...)
}
\arguments{
\item{folder}{a study or collection of variables}

\item{name}{a tableset to move}

\item{new_folder}{a subset of the studies new location}
}
\description{
Move the workspace
}
\examples{
\dontrun{
move_workspace(
  folder = "",
  name = "",
  new_folder = ""
)
}

}
