% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{armadillo.create_project}
\alias{armadillo.create_project}
\title{Create a project for a variable collection}
\usage{
armadillo.create_project(
  project_name = NULL,
  users = NULL,
  overwrite_existing = "choose"
)
}
\arguments{
\item{project_name}{The name of the project to create. The project name
\itemize{
  \item{cannot be empty.}
  \item{must be no more than 56 characters.}
  \item{cannot end with a \code{-}.}
  \item{must consist of lowercase letters and numbers.}
  }}

\item{users}{A list collection of the users that should have access to the project}

\item{overwrite_existing}{Character, specifying action to take if project still exists: 'choose'
(default) displays a menu giving the option to overwrite or not, 'yes' overwrites the
existing project and 'no' exists the function with a message.}
}
\description{
Create a project for a variable collection
}
\examples{
\dontrun{
armadillo.create_project("gecko")
}

}
