% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{armadillo.install_packages}
\alias{armadillo.install_packages}
\title{Install package}
\usage{
armadillo.install_packages(paths, profile = "default")
}
\arguments{
\item{paths}{the path(s) to the package(s), can be a vector or a string}

\item{profile}{the selected profile}
}
\description{
Installs a user defined package into the provided profile. The package
is automatically whitelisted after installation. Only available during
development.
}
