% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder.R
\name{armadillo.list_folders}
\alias{armadillo.list_folders}
\title{List all folders}
\usage{
armadillo.list_folders()
}
\value{
list all folders (user and shared)
}
\description{
List all folders
}
\examples{
\dontrun{
armadillo.list_folders()
}

}
