% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{armadillo.load_resource}
\alias{armadillo.load_resource}
\title{Load a resource from a project}
\usage{
armadillo.load_resource(project, folder, name)
}
\arguments{
\item{project}{study or collection variables}

\item{folder}{the folder containing the resource}

\item{name}{name of the resource}
}
\value{
the loaded resource
}
\description{
Load a resource from a project
}
\examples{
\dontrun{
armadillo.load_resource(
  project = "gecko",
  folder = "core_all",
  name = "lc_core_1"
)
}

}
