% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{armadillo.load_table}
\alias{armadillo.load_table}
\title{Load a table from a project}
\usage{
armadillo.load_table(project, folder, name)
}
\arguments{
\item{project}{study or collection variables}

\item{folder}{the folder containing the table}

\item{name}{name of the table}
}
\value{
the contents of the table file, as data frame
}
\description{
Load a table from a project
}
\examples{
\dontrun{
armadillo.load_table(
  project = "gecko",
  folder = "core_all",
  name = "lc_core_1"
)
}

}
