% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{armadillo.load_workspace}
\alias{armadillo.load_workspace}
\title{Load workspace based upon study folder and tableset}
\usage{
armadillo.load_workspace(folder, name, env = parent.frame())
}
\arguments{
\item{folder}{study or collection variables}

\item{name}{tableset containing the subset}

\item{env}{The environment in which you want to load the objects in the
workspace. Default is the parent.frame() from which the function is called.}
}
\value{
NULL, invisibly
}
\description{
Load workspace based upon study folder and tableset
}
\examples{
\dontrun{
armadillo.load_workspace(
  folder = "gecko",
  name = "lc_core_1"
)

armadillo.load_workspace(
  folder = "gecko",
  name = "lc_core_1",
  env = globalenv()
)
}

}
