% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{armadillo.load_workspace}
\alias{armadillo.load_workspace}
\title{Load workspace based upon study folder and tableset}
\usage{
armadillo.load_workspace(folder, name, target_env = parent.frame())
}
\arguments{
\item{folder}{study or collection variables}

\item{name}{tableset containing the subset}

\item{target_env}{the environment in which you want to load the workspaces

The default environment is `parent.frame()` which means always the calling environment (usually known as `.GlobalEnv`).}
}
\description{
Load workspace based upon study folder and tableset
}
\examples{
\dontrun{
armadillo.load_workspace(
  folder = "gecko",
  name = "lc_core_1",
  target_env = parent.frame()
)
}

}
