% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{armadillo.login}
\alias{armadillo.login}
\title{Login}
\usage{
armadillo.login(armadillo)
}
\arguments{
\item{armadillo}{URL of the Armadillo server,}
}
\value{
the id token
}
\description{
Interactively obtains an id token and uses it to create a session token for
an Armadillo Service
}
\examples{
\dontrun{
armadillo.login(
  "https://armadillo.dev.molgenis.org"
)
armadillo.login("http://localhost:8080")
}

}
