% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{armadillo.move_table}
\alias{armadillo.move_table}
\title{Move the table}
\usage{
armadillo.move_table(
  project,
  folder,
  name,
  new_project = project,
  new_folder = folder,
  new_name = name
)
}
\arguments{
\item{project}{a study or collection of variables}

\item{folder}{the folder containing the table to move}

\item{name}{a table to move}

\item{new_project}{the project to move the table to}

\item{new_folder}{the folder to move the table to, defaults to folder}

\item{new_name}{use to rename the file, defaults to name}
}
\value{
NULL, invisibly
}
\description{
Move the table
}
\examples{
\dontrun{
armadillo.move_table(
  project = "gecko",
  folder = "core_all",
  name = "table1",
  new_project = "gecko",
  new_folder = "core_all_v2",
)
}

}
