% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{armadillo.subset_definition}
\alias{armadillo.subset_definition}
\title{Builds an R object containing info required to make subsets}
\usage{
armadillo.subset_definition(reference_csv = NULL, vars = NULL)
}
\arguments{
\item{reference_csv}{\code{.csv} file containing vars to subset}

\item{vars}{Deprecated: use \code{reference_csv} instead}
}
\value{
A dataframe containing variables that is used for input in the
\code{armadillo.subset()} method
}
\description{
Builds an R object containing info required to make subsets
}
\examples{
\dontrun{
armadillo.subset_definition(
  reference_csv = "C:/tmp/vars.csv"
)
}

}
