% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{create_workspace}
\alias{create_workspace}
\title{Combines tables in an .Rdata file and uploads them to a folder.}
\usage{
create_workspace(folder, name, dataset)
}
\arguments{
\item{folder}{foldername to upload to}

\item{name}{name of the .Rdata file}

\item{dataset}{vector of tables}
}
\description{
Combines tables in an .Rdata file and uploads them to a folder.
}
\examples{
\dontrun{
create_workspace(
  folder = "gecko",
  name = "johan_subset_1",
  dataset = "table1and2"
)
}

}
