% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder.R
\name{delete_user_folder}
\alias{delete_user_folder}
\title{Deletes a user's folder from the Users bucket}
\usage{
delete_user_folder(user_name)
}
\arguments{
\item{username}{the folder to delete}
}
\description{
Deletes a user's folder from the Users bucket
}
