% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder.R
\name{.delete_bucket}
\alias{.delete_bucket}
\title{Delete bucket}
\usage{
.delete_bucket(bucket_name)
}
\arguments{
\item{bucket_name}{specify name of the bucket usually a collection of variables}
}
\description{
Delete bucket
}
\keyword{internal}
