% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{list_workspaces}
\alias{list_workspaces}
\title{List the workspaces}
\usage{
list_workspaces(folder, ...)
}
\arguments{
\item{folder}{the folder in which the workspaces are located}
}
\description{
List the workspaces
}
\examples{
\dontrun{
list_workspaces(
  folder = "gecko"
)
}

}
