% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{Login to the Armadillo service}
\usage{
login(username, password)
}
\arguments{
\item{username}{username for Armadillo service (using oAuth)}

\item{password}{password for Armadillo service (using oAuth)}
}
\description{
Login to the Armadillo service
}
