% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{login}
\alias{login}
\title{Login}
\usage{
login(armadillo, minio, duration = 900)
}
\arguments{
\item{armadillo}{URL of the Armadillo server,}

\item{minio}{URL of the Armadillo MinIO server}

\item{duration}{MinIO session duration in seconds}
}
\value{
the id token
}
\description{
Interactively obtains a token and logs in to the Armadillo MinIO server.
}
\examples{
\dontrun{
login(
  "https://armadillo.dev.molgenis.org",
  "https://armadillo-minio.dev.molgenis.org"
)
login("http://localhost:8080", "http://localhost:9000")
}

}
