% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolgenisArmadillo-package.R
\docType{package}
\name{MolgenisArmadillo-package}
\alias{MolgenisArmadillo}
\alias{MolgenisArmadillo-package}
\title{MolgenisArmadillo: Armadillo Client for the Armadillo Service}
\description{
A set of functions to manage data shared on a 'MOLGENIS Armadillo' server.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/molgenis/molgenis-r-armadillo/}
  \item \url{https://molgenis.github.io/molgenis-r-armadillo/}
  \item Report bugs at \url{https://github.com/molgenis/molgenis-r-armadillo/issues/}
}

}
\author{
\strong{Maintainer}: Mariska Slofstra \email{m.k.slofstra@umcg.nl} (\href{https://orcid.org/0000-0002-0400-0468}{ORCID})

Authors:
\itemize{
  \item Sido Haakma \email{s.haakma@rug.nl} (\href{https://orcid.org/0000-0003-1278-9144}{ORCID})
  \item Tommy de Boer \email{t.de.boer01@umcg.nl} (\href{https://orcid.org/0000-0002-4492-7225}{ORCID})
  \item Fleur Kelpin \email{fleur@kelpin.nl} (\href{https://orcid.org/0000-0003-1527-5130}{ORCID})
  \item Morris Swertz \email{m.a.swertz@rug.nl} (\href{https://orcid.org/0000-0002-0979-3401}{ORCID})
  \item Tim Cadman \email{t.j.cadman@umcg.nl} (\href{https://orcid.org/0000-0002-7682-5645}{ORCID})
}

}
\keyword{internal}
